+++ ascat/timeseries.py
@@ -188,18 +188,12 @@ class StaticLayers(object):
         fn_format = 'porosity.nc'
         self.porosity = GriddedPointData(path, grid, fn_format=fn_format)
 
-    def close(self):
-        """
-        Close all file objects.
-        """
+    def __del__(self):
         self.topo_complex.close()
         self.wetland_frac.close()
         self.frozen_prob.close()
         self.snow_prob.close()
         self.porosity.close()
-
-    def __del__(self):
-        self.close()
 
 
 class AscatNc(GriddedNcContiguousRaggedTs):
@@ -345,15 +339,6 @@ class AscatNc(GriddedNcContiguousRaggedTs):
                              wetland_frac, porosity_gldas, porosity_hwsd)
 
         return ts
-
-    def close(self):
-        """
-        Close file-like objects.
-        """
-        if self.slayer is not None:
-            self.slayer.close()
-
-        super(AscatNc, self).close()
 
 
 class AscatSsmCdr(AscatNc):